requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
pymongo==3.9.0
requests==2.22.0
dnspython==1.16.0